+++ app/home/page.tsx
@@ -1,0 +1,7 @@
+import HomeClient from "./page-client";
+
+export const dynamic = "force-dynamic";
+
+export default async function page() {
+  return <HomeClient />;
+}