+++ hooks/use-mobile.tsx
@@ -1,5 +1,6 @@
-import * as React from "react";
-import logger from "@/lib/logger";
+"use client";
+
+import React, { useEffect } from "react";
 
 const MOBILE_BREAKPOINT = 768;
 
@@ -8,22 +9,19 @@ export function useIsMobile() {
     undefined
   );
 
-  React.useEffect(() => {
+  useEffect(() => {
     const initialMobile = window.innerWidth < MOBILE_BREAKPOINT;
-    logger.debug("Initial mobile detection", { isMobile: initialMobile });
     setIsMobile(initialMobile);
 
     const mql = window.matchMedia(`(max-width: ${MOBILE_BREAKPOINT - 1}px)`);
     const onChange = () => {
       const newMobile = window.innerWidth < MOBILE_BREAKPOINT;
-      logger.debug("Media query change detected", { isMobile: newMobile });
       setIsMobile(newMobile);
     };
     mql.addEventListener("change", onChange);
 
     const handleResize = () => {
       const newMobile = window.innerWidth < MOBILE_BREAKPOINT;
-      logger.debug("Window resize detected", { isMobile: newMobile });
       setIsMobile(newMobile);
     };
     window.addEventListener("resize", handleResize);